+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Replace viridis dependency by viridisLite (@olivroy, #897)
 
+* `addRasterImage` now takes `options = gridOptions()`, so that arbitrary Leaflet layer options can be controlled. (#692)
+
 # leaflet 2.2.1
 
 * When `addProviderTiles()` is used with `{leaflet.providers}` version 2.0.0 or later, the `leaflet-providers` HTML dependency produced can be correctly cached by knitr. When used with older versions of `{leaflet.providers}`, the HTML dependency uses temp files that break knitr's caching mechanism (thanks @qdread, @jaredlander; #884).